GeneralUI: Hide ugly Compat Mode Button
If you are on a phone and have played with the "Hybrid"
options we've been adding you may notice the ugly space consuming
comptablity mode button. This hides it and may be updated to hide
more stuff as nessicary to make the hybrid experience better.

Requires the custom navbar for tabs to compile or work.

Change-Id: I77dc16c4d15854c26602a4444ce682efd1f91960